fix: configure filelog with only instrumented apps' logs (#1080)
Approach we were taking was to select only relevant files from
`/var/log/pods` for `filelog` receiver.

The files are specific to **pods**, but we don't want change the
configuration every time there are changes on the pod level, so we
wildcard the pod-specific part of the files out.

However, since there are easy collisions between statefulsets and
deployments/daemonsets we have to glob the file pattern for each of
those types **and** be specific per the kind of the instrumented
app(deployment/statefulset/daemonset).

Next-best alternative to this would be to fork `filelog` altogether and
make it very aware of those rules, keep the whole list of instrumented
app in memory and watch the `/var/log/pods` for changes.

It's not pretty, but I think it's better than what we have now - totally
broken logs pipeline. I'd continue with this general approach(happy to
take suggestions about how to implement it better), and work on a
long-term solution later. Latter probably comes with a considerably more
work.