Add env vars to runtime details (#1100)
This PR addresses #1083 

The main goal of this PR is to avoid overriding existing environment
variables with the values used for instrumentation purposes.

* Adding instrumentation-related environment variables to runtime
inspection (PYTHONPATH and NODE_OPTIONS currently).
* In case such env var is defined both in the container build and in a
k8s template the one in the template will take precedence.
* Support reverts once Odigos is uninstalled or instrumentation is
removed - we should return to the original state. For that purpose, we
added an annotation stating the original instrumentation-related env var
values from the k8s manifest if there were any. This annotation is used
to undo any env var changes we might have done, and the annotation
itself is deleted as part of uninstall/un-instrument.
* Added unit tests for the modified `Revert` and
`ApplyInstrumentationDevicesToPodTemplate` functions.

TODO: update helm charts to be consistent with this change. PR in
https://github.com/keyval-dev/odigos-charts/pull/38