feat: rename GetName() to GetID() and add debug, uptrace configurers (#1234)
Sorry about the huge rename, but I made a mistake of mirroring the
previous method name instead of thinking critical about what that did.
`GetID` is much better name because all that has to do is to provide a
reference for a destination if two of the same types are used. On the VM
Agent side we use the array index for that. On k8s we create a random
hash.

Other than that, this PR
- adds two new destination types: debug and Uptrace
- skips failing destinations instead of erroring out of compiling the
Collector config altogether.
- emits warnings when no signals are specified for a destination.

Commits:

- **feat: add debug destination type**
- **feat: add Uptrace destination type**
- **feat: continue compiling collector config if one of the exporters is
missing**
- **docs: add a comment about assumptions made for GetName**
- **feat: make debug and uptrace configers available**
- **feat: warn when no signals configured**
- **reafactor: rename GetName to GetID**