Collectors self-telemetry pipelines. (#1431)
Send collector metrics from node and cluster collectors.
The goal is to have metrics about the data sent and throughput from
sources and to destinations.

The main parts of this PR are:
* Collector configuration for collecting self-telemetry. This includes
definning the self-telemetry pipeline in the node and cluster
collectors.
* `odigostrafficmetrics` which is a processor that additional collects
metrics to those that are already provided by the collector. This
processor can be configured to attach different attributes to the
metrics. In addition, it is possible to configure the processor to
perform its measurements on a fraction of the spans/metrics/logs.
* `collectormetrics` package in the UI server which acts as an OTLP
receiver, and exposes endpoints for the UI.