fix: remove only protected c files and their directories (#1441)
During an issue with eBPF instrumentation breaking upon odiglet
restarts, I implemented a "smart" recreation of instrumentation files.
The goal was to retain compiled C files when odiglet restarted. However,
this approach led to other files not being properly recreated in certain
cases. For example, if the instrumentation version was upgraded, the
host filesystem could end up with both versions, which caused problems,
especially in Python.

This PR addresses the issue by ensuring that everything is deleted and
recreated except for the C files and their directories. For instance, in
`/var/odigos/python-ebpf/`, everything will be recreated except for
`pythonUSDT.abi3.so`.

---------

Co-authored-by: Tamir David <tamirdavid@Tamirs-MacBook-Pro.local>