feat: remove OdigosConfig CRD (#1405)
This will remove usage of `odigosconfigurations.odigos.io` CRD and
replace it with a plain ConfigMap to remove race conditions when
installing named CRD and CR at the same time during the install process.
New ConfigMap will look as following:

```
[...]
data:
  config.yaml: { ... serialized JSON of odigosconfiguration.spec ... }
[...]
```

With the exception of removal of `SupportedSDKs`(as per @blumamir
request to remove it from CM) the contents of `config.yaml` value and
the `Spec` field are identical and serialize from the same string and
parse to the same object.

Installs now just install ConfigMap instead of the CR of
OdigosConfiguration, Helm upgrades are unchanged, as they used to just
overwrite the CR anyway. CLI parses the old config unless there's
already a ConfigMap and creates it as a ConfigMap.

The old CRD is left there unused for now to be able to keep on using the
API for it(for migration) - can be removed when we feel we are ready to
remove the migration path.