test: add e2e test for odigos cli upgrade from latest to HEAD (#1446)
Since odigos upgrade is a potentially sensative process which can fail
for various reasons, this PR adds an e2e test that starts with the
latest odigos version from github, and then calls cli upgrade to the
current version in HEAD.

Then it generates traffic and verify that traces are created.

- This is a basic test for the most straight forward case. while it can
catch many issues, we can consider extending it in the future to run and
assert more aspects after the upgrade
- It currently only runs cli upgrade and not helm upgrade. since we are
soon going to migrate cli to helm as well, I think it can be taken care
of after the refactor. @rauno56
- I duplicated many of the existing yamls from existing tests, with
slight modifications. looks like we can extract some of it to a common
directory and reuse it?