refactor: cleanups for k8sutils workload package. create enums and move common functions (#1449)
This is another step in the way to move everything related to common
things into the common packages, and avoid duplications and untyped
usages in the project.

This PR:
- adds more files to the `workload` package in k8sutils to differentiate
between the different utils and enums
- enums for the workload kinds
- function to handle and convert workload kinds
- rename function to common format
- replace usages in controllers code to use enums where possible

There are countless places to go over and improve. This PR is touching
on the most easy and straight forward cases I spotted.