feat: add odigos describe analyze phase to make the info accessible to tools (#1633)
Until this PR, everytime we calculated the `odigos describe` output, we
transformed it to text, which makes it less useful to be used in odigos
ui and other tools, as the data is human readable and not machine
readable.

This PR separates the 2 tasks. Now one can simply get the results of the
`odigos describe` command and now immediately if something is still in
transit or has errors.
Tools can process this data to display it nicely or act upon it. This
function offers an handy aggregation for the relevant data in the
context of odigos entities.