Create a common predicate for contaniers ready, use it in Pod reconciler for instrumentation (#1754)
Following #1666, this PR includes the following:

1. Change the predicate for ebpf instrumentation to only pass events
where all the containers in a pod become ready. This will fix cases
where the reconciliation was called too early and missed the relevant
process.
2. Create an `AllContainersReadyPredicate` which is common to the
runtime details reconciler (the predicate was added in #1666 ) and the
ebpf instrumentation reconciler.
3. Create a common `DeletionPredicate` predicate - for the
instrumentation reconcilation it is an OR between the containers ready
predicate and the delete one.