[GEN-1784]: apply icons for "monitors" in drawer card-details (#1882)
This pull request includes several changes to the
`frontend/webapp/components` directory, focusing on refactoring
components and improving the styling and functionality of the
`CardDetails`, `ConfiguredFields`, and `MonitorsLegend` components.

Refactoring and Interface Updates:

*
[`frontend/webapp/components/common/card-details/index.tsx`](diffhunk://#diff-a1aeb311b15066559d3a82bef324c454a7de3a09e06ab8637b9b36c2a79e655dL3-R6):
Renamed `CardDetailsProps` to `Props` and moved the `ConfiguredFields`
import to the top of the file. The `CardDetails` component was
restructured and re-exported at the bottom of the file.
[[1]](diffhunk://#diff-a1aeb311b15066559d3a82bef324c454a7de3a09e06ab8637b9b36c2a79e655dL3-R6)
[[2]](diffhunk://#diff-a1aeb311b15066559d3a82bef324c454a7de3a09e06ab8637b9b36c2a79e655dL14-L26)
[[3]](diffhunk://#diff-a1aeb311b15066559d3a82bef324c454a7de3a09e06ab8637b9b36c2a79e655dR28-R38)
*
[`frontend/webapp/components/common/configured-fields/index.tsx`](diffhunk://#diff-03463bf02731c500640be42702debbd088df7a64298eac7d7474f5864f0c6022L4-R15):
Renamed `TypeDetail` to `Detail` and `ConfiguredFieldsProps` to `Props`.
Added a new `renderValue` function to handle different types of values.
[[1]](diffhunk://#diff-03463bf02731c500640be42702debbd088df7a64298eac7d7474f5864f0c6022L4-R15)
[[2]](diffhunk://#diff-03463bf02731c500640be42702debbd088df7a64298eac7d7474f5864f0c6022L36-R37)
[[3]](diffhunk://#diff-03463bf02731c500640be42702debbd088df7a64298eac7d7474f5864f0c6022L47-R49)
[[4]](diffhunk://#diff-03463bf02731c500640be42702debbd088df7a64298eac7d7474f5864f0c6022R79-R97)
*
[`frontend/webapp/components/overview/monitors-legend/index.tsx`](diffhunk://#diff-a4d5b4e054905b04e1d8a5d67e6844925c981e65138fe5be7cec01cfe42ba2dcL3-L43):
Added a `Props` interface and updated the `MonitorsLegend` component to
use the new props. The component now filters signals and adjusts styles
based on the provided props.

Styling Improvements:

*
[`frontend/webapp/components/common/configured-fields/index.tsx`](diffhunk://#diff-03463bf02731c500640be42702debbd088df7a64298eac7d7474f5864f0c6022L25-R26):
Updated the `ItemTitle` component to use theme-based coloring.
*
[`frontend/webapp/components/overview/monitors-legend/index.tsx`](diffhunk://#diff-a4d5b4e054905b04e1d8a5d67e6844925c981e65138fe5be7cec01cfe42ba2dcL3-L43):
Updated `List`, `ListItem`, and `MonitorTitle` components to use dynamic
styling based on the new props.