[GEN-1845]: add "withLabels" prop to monitor icons component (#1899)
This pull request includes changes to the `MonitorsIcons` component and
its usage in the `DataCard` component. The main updates involve adding
labels to monitor icons and adjusting the icon size and spacing.

Changes to `MonitorsIcons` component:

*
[`frontend/webapp/reuseable-components/monitors-icons/index.tsx`](diffhunk://#diff-ae9f732b524d8587d16510849079e64475b475080cdb810b6f38e52108731d81L5-R26):
Added a new `withLabels` prop to optionally display labels next to the
monitor icons and adjusted the gap between icons.

Changes to `DataCard` component:

*
[`frontend/webapp/reuseable-components/data-card/data-card-fields/index.tsx`](diffhunk://#diff-ce146776b402d3e7c41f2ebb82b32a2b8636ec73200d70a6f30f0f08dc1da6d0L77-R77):
Updated the `MONITORS` case to use the new `withLabels` prop and changed
the icon size from 14 to 12.