[GEN-1927]: Cypress test for onboarding-flow (#1930) (#1931)
This pull request includes several changes aimed at improving the
frontend codebase and updating the Cypress end-to-end testing
configuration. The most important changes include adding unique IDs to
various components, updating Cypress configuration and tests, and
removing an outdated script.

Improvements to frontend components:

*
[`frontend/webapp/containers/main/destinations/destination-modal/choose-destination-body/destinations-list/destination-list-item/index.tsx`](diffhunk://#diff-0111152ee1a677294919adfc934717a9e3f9d1cfa99362d49660ed218f12ae18L95-R95):
Added unique IDs to `ListItem` components for better identification.
*
[`frontend/webapp/containers/main/destinations/destination-modal/choose-destination-body/destinations-list/index.tsx`](diffhunk://#diff-0fc3893b30022c16fec91cb64d38fa36de0dc4e8eabbbbb7a3b94701b45a4badL52-R52):
Updated the key for `DestinationListItem` to use `categoryItem.type`
instead of `categoryItem.displayName`.
*
[`frontend/webapp/containers/main/sources/choose-sources/choose-sources-body/choose-sources-body-fast/sources-list/index.tsx`](diffhunk://#diff-6b67649d370d208941f4e5a78c0c9de2f5b9b65fd49e0ede8ef57066982450aaL139-R139):
Added unique IDs to `Group` components for better identification.
*
[`frontend/webapp/reuseable-components/no-data-found/index.tsx`](diffhunk://#diff-6807c50cee51d35309adcaaab02fdb295cdde0392d6d0e010c0e602a21ae244fL37-R37):
Added an ID to the `Container` component for better identification.

Updates to Cypress configuration and tests:

*
[`frontend/webapp/cypress.config.ts`](diffhunk://#diff-324fc1c28a87dedafd74a9513d1c38b2ab26e47a3c5e9a9a9e1b74dce75e0aa2R3-R9):
Set the `baseUrl` to use a variable for easier configuration.
*
[`frontend/webapp/cypress/e2e/connection.cy.ts`](diffhunk://#diff-dfc1066bf4c3d7d672402ac7c804872b2500bcdb7f9c1736628b4a7e710b0647R1-R10):
Added a new test to verify the frontend and backend connections using
GraphQL.
*
[`frontend/webapp/cypress/e2e/onboarding.cy.ts`](diffhunk://#diff-709b703b03f6676250de6b700a4d517e5d5622ad788d1f574aa2489527d77fc7L2-R26):
Updated onboarding tests to include checks for the existence of a
"default" namespace and the "Jaeger" destination.

Removal of outdated script:

*
[`tests/common/odigos_ui.sh`](diffhunk://#diff-16f3e24ddc9ddd226f571aea7b04628df0c0917d6bf2577465b0b2654c94e124L1-L125):
Removed the old script for managing the Odigos UI.

Updates to package scripts:

*
[`frontend/webapp/package.json`](diffhunk://#diff-ccf6337b0064354343f900ffd8b4ee91aa8cd014f3292548a800ad3dac39c1f4L6-R10):
Updated Cypress-related scripts and removed unnecessary build scripts.
[[1]](diffhunk://#diff-ccf6337b0064354343f900ffd8b4ee91aa8cd014f3292548a800ad3dac39c1f4L6-R10)
[[2]](diffhunk://#diff-ccf6337b0064354343f900ffd8b4ee91aa8cd014f3292548a800ad3dac39c1f4L42-R38)

Changes to end-to-end test configuration:

*
[`tests/e2e/ui/chainsaw-test.yaml`](diffhunk://#diff-755f1b6537d52a2a2e4be3c1e17d296567f0b9e43c396b3c552a8dff9851a2bbL4-R56):
Updated the test configuration to use the Odigos CLI for installing and
starting the UI, and to run Cypress tests.