[GEN-1924]: add CRUD E2E tests for UI sources (#1958)
This pull request includes several changes to improve the functionality
and user experience of the frontend web application. The most important
changes include adjustments to the notification toast positioning,
enhancements to the multi-source control component, updates to the
source body form, new end-to-end tests for source CRUD operations, and
improvements to the `useSourceCRUD` hook.

### UI/UX Improvements:
*
[`frontend/webapp/components/notification/toast-list.tsx`](diffhunk://#diff-c6ead7587d1e5c52295c921dde2a1a5b9b06977ab3900c7dbf475c47cc1c664cL10-R10):
Changed the position of the notification toast from the bottom to the
top of the screen.
*
[`frontend/webapp/containers/main/overview/multi-source-control/index.tsx`](diffhunk://#diff-cdd99a8fcd0484b24586f22b1f7b1bcf8d964bead53eeb0b2c07646c7301af70L60-R60):
Added a `data-id` attribute to the `Transition` component for better
identification.
*
[`frontend/webapp/containers/main/sources/update-source-body/index.tsx`](diffhunk://#diff-c5285ffb349464d72b44f6c535fbe1a02959ef2213441d34027c3fa579207565R25):
Added a `name` attribute to the input element for the source name to
improve form handling.

### Testing Enhancements:
*
[`frontend/webapp/cypress/e2e/03-sources.cy.ts`](diffhunk://#diff-981488421b684f7ca98e7e8b147b14c23fe2c57e876d88ca981c563c8ba003cdR1-R105):
Added new Cypress end-to-end tests for creating, updating, and deleting
CRDs in the cluster.

### Codebase Improvements:
*
[`frontend/webapp/hooks/sources/useSourceCRUD.ts`](diffhunk://#diff-1be86445dc0ed6bfdc08f7525c5800ef39bdcd70318a3c130f83f36f1ae9f5c8L43-R50):
Improved error handling and variable access by adding optional chaining
to the `sources` array.