[GEN-1986]: fix values for destination dropwdowns (#1971)
This pull request includes a modification to the
`DestinationDynamicFields` component in the
`frontend/webapp/containers/main/destinations/destination-form-body/dynamic-fields/index.tsx`
file. The change updates the handling of the `Dropdown` input type to
ensure proper value assignment and selection.

Changes to `DestinationDynamicFields` component:

*
[`frontend/webapp/containers/main/destinations/destination-form-body/dynamic-fields/index.tsx`](diffhunk://#diff-c17889e1ffea148a206f57df5da9f8cc755e7e522a9057a339f6887c01ed5815L19-R19):
Updated the `Dropdown` input type to set the `value` prop correctly and
simplified the `onSelect` callback to pass only the selected option's
value.