Reload Collectors without restart when ConfigMap changes (#1903)
Kubernetes automatically updates mounted configmaps in a pod (see
https://kubernetes.io/docs/tutorials/configuration/updating-configuration-via-a-configmap/).
But, the workload needs to watch for the update itself. So in our case,
if the collector knows to watch for file changes, it can automatically
signal a dynamic config reload without restarting the collector.

This forks the default fileprovider in our collectors to create a new
odigosfileprovider that uses fsnotify to watch for updates to the
collector ConfigMap. When the ConfigMap is updated, the new fileprovider
will signal the collector to hot-reload its config. Technically, we
watch for `fsnotify.Remove` events, because the projected configmap data
is a symlink, not an actual copy of the configmap (meaning that watching
`fsnotify.Write` doesn't trigger any update).

This means we don't need to restart the collector deployments or
daemonsets for basic config updates, so those controllers have been
updated to no longer update deployments when just the configmap has
changed. They can of course still be manually redeployed with `kubectl`.

In my manual testing, it took about 1 minute for the change to be
reflected, which is due to the default kubelet sync period
(https://kubernetes.io/docs/tasks/configure-pod-container/configure-pod-configmap/#mounted-configmaps-are-updated-automatically).
Not sure how we could test for this automatically but working on that

---------

Co-authored-by: Ben Elferink <ben.elferink@icloud.com>