[GEN-1837]: "other agent" for new UI (#1994)
This pull request includes several changes focused on improving the user
interface and functionality of the `frontend/webapp` components. The
most important changes include refactoring the
`ConfiguredDestinationsList` component, enhancing the `DataTab`
component, and adding new features to handle specific conditions in the
`SourceDrawer`.

### Refactoring and UI Improvements:

*
[`frontend/webapp/containers/main/destinations/add-destination/configured-destinations-list/index.tsx`](diffhunk://#diff-f5288a6a2e6ccce16445bdd7486abda1bbd12662ea0e0bc846ff0f60b85b12f9R4-R8):
Refactored `ConfiguredDestinationsListItem` to use `DataTab` instead of
custom-styled components, simplifying the code and improving
maintainability.
[[1]](diffhunk://#diff-f5288a6a2e6ccce16445bdd7486abda1bbd12662ea0e0bc846ff0f60b85b12f9R4-R8)
[[2]](diffhunk://#diff-f5288a6a2e6ccce16445bdd7486abda1bbd12662ea0e0bc846ff0f60b85b12f9L21-R41)
[[3]](diffhunk://#diff-f5288a6a2e6ccce16445bdd7486abda1bbd12662ea0e0bc846ff0f60b85b12f9L160-R59)

### Enhancements to `DataTab` Component:

*
[`frontend/webapp/reuseable-components/data-tab/index.tsx`](diffhunk://#diff-afbf1606ee29d25bd9273d63b66a4746dbdd1cd8406ad651cfd123ac28f72006L1-L21):
Enhanced the `DataTab` component by adding support for extended content,
actions, and monitor labels. This change improves the flexibility and
usability of the component.
[[1]](diffhunk://#diff-afbf1606ee29d25bd9273d63b66a4746dbdd1cd8406ad651cfd123ac28f72006L1-L21)
[[2]](diffhunk://#diff-afbf1606ee29d25bd9273d63b66a4746dbdd1cd8406ad651cfd123ac28f72006R58)
[[3]](diffhunk://#diff-afbf1606ee29d25bd9273d63b66a4746dbdd1cd8406ad651cfd123ac28f72006L74-R144)

### Handling Specific Conditions in `SourceDrawer`:

*
[`frontend/webapp/containers/main/sources/source-drawer-container/index.tsx`](diffhunk://#diff-d4745da2f63be90cb1f7af8ea29fb2f37cf0a80f0f04a207b6fc9149855ca6d5L11-R11):
Added logic to detect the presence of other agents in `SourceDrawer` and
updated the container details accordingly. This change helps in
providing more accurate information to the user.
[[1]](diffhunk://#diff-d4745da2f63be90cb1f7af8ea29fb2f37cf0a80f0f04a207b6fc9149855ca6d5L11-R11)
[[2]](diffhunk://#diff-d4745da2f63be90cb1f7af8ea29fb2f37cf0a80f0f04a207b6fc9149855ca6d5R82-R92)

### Additional UI Adjustments:

*
[`frontend/webapp/reuseable-components/condition-details/index.tsx`](diffhunk://#diff-8ddad4625b70db190b926980070c30fb0c429fb2c13be252bba2c6de1e8cef4dL63-R63):
Adjusted the alignment of the `ExtendIcon` in the `ConditionDetails`
component to improve visual consistency.
*
[`frontend/webapp/reuseable-components/extend-icon/index.tsx`](diffhunk://#diff-152d17ee6b8a621b34f3390f999aa5d72b0e739cbaee55a24022c91a1389774eL8-R22):
Updated the `ExtendIcon` component to support center alignment,
enhancing its flexibility in various layouts.

These changes collectively enhance the user experience and
maintainability of the codebase.