Update doc to new UI (#1993)
This pull request includes a change to the
`docs/quickstart/installation.mdx` file in the `odigos ui`
documentation. The change updates the instructions for accessing the
legacy UI instead of the experimental new UI.

Documentation update:

*
[`docs/quickstart/installation.mdx`](diffhunk://#diff-4f768f9645352fa629b2c955f5d407945d5000f12746fb50929cc70b599350f2L79-R87):
Changed the section from "Experimental: New UI with --beta flag" to
"Using the Legacy UI" and updated the corresponding command and warning
message.