feat: helm legacy UI flag (#2009)
This pull request includes changes to update the naming convention for
the UI ports from "beta-ui" to "legacy-ui" across multiple files. The
most important changes are:

UI Port Naming Updates:

*
[`cli/cmd/resources/ui.go`](diffhunk://#diff-c286e10d34710a80a59127b2b7951e8a33d9b9554e47d2f2b827fd690f2e53abL317-R317):
Changed the port name from "beta-ui" to "legacy-ui" in the
`NewUIService` function.
*
[`cli/cmd/ui.go`](diffhunk://#diff-bdc4ded2bf4628241563326d5d7b0443b57d2d37b674014da632b14b78a1c309L166-R166):
Updated the flag description from "beta" to "legacy" in the `init`
function to reflect the new naming convention.
*
[`helm/odigos/templates/ui/service.yaml`](diffhunk://#diff-042cac3e8292ac2b7d917f5012c9d17350361c1f23c6d08252c4438580367776L12-R12):
Renamed the port from "beta-ui" to "legacy-ui" in the `spec` section.

---------

Co-authored-by: alonkeyval <alonbraymok007@gmail.com>