Track un-active instrumentation for better reporting (#2016)
This PR changes when the instrumentation manager tracks an
instrumentation.
Once an instrumentation is initialized, we should start tracking it even
if it fails to load.
The reason is, that we need to call the reporter once the process exits
(so it can report/clean it) - however, if we only track the PID once the
instrumentation is loaded we won't be able to invoke the reporter
properly once a failed-to-load instrumentation needs to be cleaned.

Each PID we track is marked whether the instrumentation is active
(loaded successfully) or not - by the inst field being nil or not.
For un-active PIDs we won't apply config updates.

In addition, in case we get an exec event on a non-active PID, we try to
instrument it again. This is helpful in cases of chain-loading where the
first executable is written in a language we can't instrument, while the
second is valid for instrumentation.