chore: narrow down the rbac permissions for schedualer (#2024)
In this PR:

- move destinations and collectorsgroup rbac permission to be role (on
odigos ns) instead of clusterrole. These live anyway only in our ns, but
it looks better in rbac reviews.
- removed unused create/delete/update/patch for the destinations, as we
only read them.
- currently the leader election role brings in configmaps permissions,
but I also added the permissions we need for the reconcilers so not to
mix unrelated things. we will need to also understand and address the
leader election role later on.
- removed the finalizers permissions, as we don't use finalizers and
it's unused.
- created some consts in cli files for better structure to the
references
- synced helm files with the changes
- tested both helm and cli locally to make sure it works

InstrumentationConfig is left as a clusterrole, since they belong to the
various namespaces where the sources reside.