[GEN-2066]: remove null values from describe (#2026)
This pull request introduces a new utility function `removeNoValues` and
integrates it into the `Code` component to improve JSON handling. The
most important changes include the addition of the `removeNoValues`
function, its export, and its use in the `Code` component with `useMemo`
to optimize performance.

### New utility function:

*
[`frontend/webapp/utils/functions/formatters/remove-no-values/index.ts`](diffhunk://#diff-191ef4d6be5f136e4f53a28f4529d78c89e4725117729142df08256267ae58d9R1-R24):
Added the `removeNoValues` function to remove null, undefined, and empty
string values from objects and arrays.

### Integration of `removeNoValues`:

*
[`frontend/webapp/reuseable-components/code/index.tsx`](diffhunk://#diff-b2d854039ed7f8b3ce6c2926bf9b37b7a766e688bd6d66819db9488355bd31ecR1-R4):
Imported `removeNoValues` and used it within the `Code` component to
clean JSON data before flattening and stringifying it. This change also
includes the use of `useMemo` to optimize performance.
[[1]](diffhunk://#diff-b2d854039ed7f8b3ce6c2926bf9b37b7a766e688bd6d66819db9488355bd31ecR1-R4)
[[2]](diffhunk://#diff-b2d854039ed7f8b3ce6c2926bf9b37b7a766e688bd6d66819db9488355bd31ecL19-R30)
*
[`frontend/webapp/utils/functions/formatters/index.ts`](diffhunk://#diff-801ac659dba2f09a5dbfcf41f6d6479ab5719cbbcea26548a42f67683ebf06faR8):
Exported the `removeNoValues` function to make it available for use in
other modules.