[GEN-2063]: open expanded view when selecting all sources (#2029)
This pull request includes a small change to the
`frontend/webapp/hooks/sources/useSourceFormData.ts` file. The change
comments out a line of code that sets the selected namespace, providing
an alternative approach if the goal is to select all without opening the
expanded view.

Code change:

*
[`frontend/webapp/hooks/sources/useSourceFormData.ts`](diffhunk://#diff-cd2bc3c9473adfb41833260ef193384279ed75ef4b6c6866290ca78a174471e7L125-R128):
Commented out the line that sets the selected namespace and added a note
explaining an alternative approach.